fix org details after rebase/file move
