+++ frontend/controller/access/organizations/OrganizationPage/OrganizationDetails.tsx
@@ -1,18 +1,55 @@
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
+import { LabelGroup, Label } from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, useParams, Link } from 'react-router-dom';
 import { PageDetail, PageDetails, SinceCell } from '../../../../../framework';
+import { useGet } from '../../../../common/useItem';
 import { RouteE } from '../../../../Routes';
 import { Organization } from '../../../interfaces/Organization';
+import { Credential } from '../../../interfaces/Credential';
+import { InstanceGroup } from '../../../interfaces/InstanceGroup';
+import { CredentialLabel } from '../../../common/CredentialLabel';
+import { ExecutionEnvironmentDetail } from '../../../common/ExecutionEnvironmentDetail';
+
+function useGalaxyCredentials(orgId: string) {
+  const { data } = useGet<{ results: Credential[] }>(
+    `/api/v2/organizations/${orgId}/galaxy_credentials/`
+  );
+  return data?.results ?? [];
+}
+
+function useInstanceGroups(orgId: string) {
+  const { data } = useGet<{ results: InstanceGroup[] }>(
+    `/api/v2/organizations/${orgId}/instance_groups/`
+  );
+  return data?.results ?? [];
+}
 
 export function OrganizationDetails(props: { organization: Organization }) {
   const { t } = useTranslation();
   const { organization } = props;
   const history = useNavigate();
+  const params = useParams<{ id: string }>();
+
+  const galaxyCredentials = useGalaxyCredentials(params.id || '0');
+  const instanceGroups = useInstanceGroups(params.id || '0');
+
+  // TODO look up license type from context (TBD) and add max hosts
   return (
     <PageDetails>
       <PageDetail label={t('Name')}>{organization.name}</PageDetail>
       <PageDetail label={t('Description')}>{organization.description}</PageDetail>
+      {/* {license_info?.license_type !== 'open' ? (
+        <PageDetail label={t('Max hosts')}>{organization.max_hosts}</PageDetail>
+      ) : null} */}
+      <ExecutionEnvironmentDetail
+        virtualEnvironment={organization.custom_virtualenv || undefined}
+        executionEnvironment={organization.summary_fields?.default_environment}
+        isDefaultEnvironment
+        helpText={t`The execution environment that will be used for jobs
+          inside of this organization. This will be used a fallback when
+          an execution environment has not been explicitly assigned at the
+          project, job template or workflow level.`}
+      />
       <PageDetail label={t('Created')}>
         <SinceCell
           value={organization.created}
@@ -40,6 +77,28 @@ export function OrganizationDetails(props: { organization: Organization }) {
             )
           }
         />
+      </PageDetail>
+      <PageDetail
+        label={t`Instance Groups`}
+        helpText={t`The Instance Groups for this Organization to run on.`}
+        isEmpty={instanceGroups.length === 0}
+      >
+        <LabelGroup>
+          {instanceGroups.map((ig) => (
+            <Label color="blue" key={ig.id}>
+              <Link to={RouteE.InstanceGroupDetails.replace(':id', (ig.id ?? 0).toString())}>
+                {ig.name}
+              </Link>
+            </Label>
+          ))}
+        </LabelGroup>
+      </PageDetail>
+      <PageDetail label={t('Galaxy Credentials')} isEmpty={galaxyCredentials.length === 0}>
+        <LabelGroup>
+          {galaxyCredentials?.map((credential) => (
+            <CredentialLabel credential={credential} key={credential.id} />
+          ))}
+        </LabelGroup>
       </PageDetail>
     </PageDetails>
   );
